+++ packages/extensions/core/lib/pipeline/plugins/transformer/index.ts
@@ -1,0 +1,1 @@
+export * from "./transformer";
